backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-huggingface
sentence-transformers
chromadb
pypdf
python-dotenv
